feat: add order and orderItem entities; create orders during checkout

+++ src/main/java/com/store/local_store/domain/ports/repos/ProductRepository.java
@@ -16,5 +16,5 @@ public interface ProductRepository {
 
     void saveAll(List<Product> productsToUpdate);
 
-    Integer updateProductStock(Integer quantity, Long productId);
+    Integer updateProductStock(Long quantity, Long productId);
 }
